+++ packages/preview/web/src/index.ts
@@ -1,6 +1,5 @@
 import { interactionMap } from './Interactions/interactionMap';
 import { defaultTransformation } from './Components/default-transformation';
-export { default as ContentStateTransformation } from './RuleEngine/ContentStateTransformation';
 export { default as RichContentPreview } from './Components/RichContentPreview';
 export { interactionMap };